# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nika Web Services",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1"
        }
    ],
    "paths": {
...
        "/cities": {
            "get": {
                "tags": [
                    "شهر"
                ],
                "summary": "لیست شهر ها بصورت صفحه بندی",
                "description": "",
                "operationId": "a7ec9e1ae16c0e72817824b283c3cb6d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "شماره صفحه",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": 1
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "تعداد نمایش در هر صفحه",
                        "schema": {
                            "type": "number"
                        },
                        "example": 10
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "جستجوی نام",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "province",
                        "in": "query",
                        "description": "جستجوی نام استان",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "شهر"
                ],
                "summary": "ثبت شهر",
                "description": "",
                "operationId": "0e59a8b17ea4e77383a49e1a2418915b",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "province_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "نام",
                                        "type": "string"
                                    },
                                    "province_id": {
                                        "description": "شناسه استان",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ثبت با موفقیت انجام شد",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/cities/{id}": {
            "post": {
                "tags": [
                    "شهر"
                ],
                "summary": "بروزرسانی شهر",
                "description": "",
                "operationId": "971b457e25297bdf6b19ec82f31e0eb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه شهر",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "_method",
                                    "name",
                                    "province_id"
                                ],
                                "properties": {
                                    "_method": {
                                        "description": "این مقدار باید بصورت ثابت شود",
                                        "type": "string",
                                        "default": "put",
                                        "enum": [
                                            "put"
                                        ]
                                    },
                                    "name": {
                                        "description": "نام",
                                        "type": "string"
                                    },
                                    "province_id": {
                                        "description": "شناسه استان",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "شهر"
                ],
                "summary": "حذف شهر",
                "description": "",
                "operationId": "6bdf73c70d6ef45540ae8c9b54c5372b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه شهر",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/grades": {
            "get": {
                "tags": [
                    "مقاطع تحصیلی"
                ],
                "summary": "لیست مقاطع تحصیلی بصورت صفحه بندی",
                "description": "",
                "operationId": "939d5b2cfb66bc732c2fac28dced4706",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "شماره صفحه",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "example": 1
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "تعداد نمایش در هر صفحه",
                        "schema": {
                            "type": "number"
                        },
                        "example": 10
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "جستجوی نام",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "مقاطع تحصیلی"
                ],
                "summary": "ثبت مقطع تحصیلی",
                "description": "",
                "operationId": "9e03a90e491c68ed9f15e99e3ea7ba0b",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "نام",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ثبت با موفقیت انجام شد",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/grades/{id}": {
            "post": {
                "tags": [
                    "مقاطع تحصیلی"
                ],
                "summary": "بروزرسانی مقطع تحصیلی",
                "description": "",
                "operationId": "e758c15e90d9b71c3da2fedc1e87b7ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه مقطع تحصیلی",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "_method",
                                    "name"
                                ],
                                "properties": {
                                    "_method": {
                                        "description": "این مقدار باید بصورت ثابت شود",
                                        "type": "string",
                                        "default": "put",
                                        "enum": [
                                            "put"
                                        ]
                                    },
                                    "name": {
                                        "description": "نام",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "مقاطع تحصیلی"
                ],
                "summary": "حذف مقطع تحصیلی",
                "description": "",
                "operationId": "f51ba9123b9db6c4cb06436b6fc8d767",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه مقطع تحصیلی",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "عملیات موفق",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "توکن را به همراه Bearer وارد کنید.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}